Minor adjustments

+++ game/ecs/v5/scene_test.go
@@ -9,7 +9,7 @@ import (
 
 var _ = Describe("Scene", func() {
 	type Position struct {
-		X, Y float64
+		X, Y int
 	}
 	type Age struct {
 		Value int
